Add ignored events tables in QA dashboard
